Extract payment methods
